+++ Lead-Server-Source/db/src/ClientManagerPlayer.cpp
@@ -97,7 +97,7 @@ size_t CreatePlayerSaveQuery(char * pszQuery, size_t querySize, TPlayerTable * p
 			"ht = %d, "
 			"dx = %d, "
 			"iq = %d, "
-			"gold = %d, "
+			"gold = %lld, "
 			"exp = %u, "
 			"stat_point = %d, "
 			"skill_point = %d, "
@@ -259,7 +259,7 @@ void CClientManager::QUERY_PLAYER_LOAD(CPeer * peer, DWORD dwHandle, TPlayerLoad
 
 		TItemCacheSet * pSet = GetItemCacheSet(pTab->id);
 
-		sys_log(0, "[PLAYER_LOAD] ID %s pid %d gold %d ", pTab->name, pTab->id, pTab->gold);
+		sys_log(0, "[PLAYER_LOAD] ID %s pid %d gold %lld ", pTab->name, pTab->id, static_cast<long long>(pTab->gold));
 
 		//--------------------------------------------
 		// item & AFFECT & QUEST loading : 
@@ -842,17 +842,17 @@ void CClientManager::__QUERY_PLAYER_CREATE(CPeer *peer, DWORD dwHandle, TPlayerC
 			"skill_level, quickslot) "
 			"VALUES(0, %u, '%s', %d, %d, %d, %d, %d, "
 			"%d, %d, %d, %d, %d, %d, %d, "
-			"%d, %d, %d, %d, %d, %d, %d, 0, %d, 0, ",
+			"%d, %d, %d, %d, %d, %d, %d, 0, %lld, 0, ",
 			GetTablePostfix(),
 			packet->account_id, packet->player_table.name, packet->player_table.level, packet->player_table.st, packet->player_table.ht, packet->player_table.dx, packet->player_table.iq,
 			packet->player_table.job, packet->player_table.voice, packet->player_table.dir, packet->player_table.x, packet->player_table.y, packet->player_table.z,
 			packet->player_table.hp, packet->player_table.sp, packet->player_table.sRandomHP, packet->player_table.sRandomSP, packet->player_table.stat_point, packet->player_table.stamina, packet->player_table.part_base, packet->player_table.part_base, packet->player_table.gold);
 
-	sys_log(0, "PlayerCreate accountid %d name %s level %d gold %d, st %d ht %d job %d",
+	sys_log(0, "PlayerCreate accountid %d name %s level %d gold %lld, st %d ht %d job %d",
 			packet->account_id, 
 			packet->player_table.name, 
 			packet->player_table.level, 
-			packet->player_table.gold, 
+			static_cast<long long>(packet->player_table.gold), 
 			packet->player_table.st, 
 			packet->player_table.ht, 
 			packet->player_table.job);
@@ -1238,4 +1238,3 @@ void CClientManager::FlushPlayerCacheSet(DWORD pid)
 		delete c; 
 	}
 }
-
